requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.10.0
astunparse==1.6.3
cachetools==6.2.4
certifi==2025.8.3
charset-normalizer==3.4.3
distlib==0.4.0
distro==1.9.0
Django==5.2.7
exceptiongroup==1.3.1
filelock==3.19.1
flatbuffers==25.2.10
gast==0.6.0
google-auth==2.45.0
google-genai==1.56.0
google-pasta==0.2.0
grpcio==1.74.0
h11==0.16.0
h5py==3.14.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
joblib==1.5.2
keras==3.11.3
libclang==18.1.1
Markdown==3.8.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.3
namex==0.1.0
numpy==2.2.6
opencv-python==4.12.0.88
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pillow==11.3.0
platformdirs==4.4.0
protobuf==6.32.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
requests==2.32.5
rich==14.1.0
rsa==4.9.1
scikit-learn==1.7.1
scipy==1.15.3
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
threadpoolctl==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
virtualenv==20.34.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.3
